Provide a 'symbolLocation' thrift method
Summary:
The most common reason people are using Glass in ad hoc ways is to find symbol locations (i.e. a mixture of search or our symbol id builders + describe().

describe() is still more than we need for this, and I keep forgetting we have 'resolveSymbolRange()'.

So make it clearer by calling it `symbolLocaton` and then I'll go through and fix a ton of the clients to use this much cheaper method. Then expose it to the main clients properly and refactor clients of describe()

Reviewed By: simonmar

Differential Revision: D59952037

fbshipit-source-id: ef7b4678995056a37d2bebda666418075c2d6381