Flag to let scip encoder infer language based on suffix
Summary:
SCIP is a multi language schema, and the language of a symbol is associated with the file the definition is contained in.
In some indexers, this is set in the scip data, and for older indexers that don't provide a scip.Language we can use --language to enforce a language for the entire indexing run.

However, for Java and Kotlin, this doesn't work, as they are intermingled in the same build ,and sometimes the same target.
So we have to fall bac.  to detecting by suffix in this case.

Language selection for symbols when converting SCIP to Glean is thus:
- if the scip.Language value is set, use it
- if --infer-language is set, attempt to use the file suffix
- otherwise use --language
- else UnknownLanguage

Reviewed By: simonmar

Differential Revision: D60156776

fbshipit-source-id: c0a94259294207c513ef123d2d7b79236f9cad37