Flat All statement
Summary:
A flattened representation for 'all' patterns. It would have been natural to have 'all' as a form of generator. But we cannot flatten the statements out of the 'all' construct, because it needs to run them to completion and collect all the results), and therefore generators would need to contains statements and that messes things up quite a lot.

Instead, this diff introduces a new form of statement, essentially on the form `X = all(p)`, for some pattern `p`. This turns out to work quite a bit better.

Reviewed By: simonmar

Differential Revision: D60121050

fbshipit-source-id: 520cb1fc954fe0d021b03c315d054fe099f30fc8