move xlang tests to new test target
Summary:
The clang codemarkup tests  (//glean/lang/codemarkup/tests/clang:tests) are very slow to run (~10mn).

To speed tests, this diff extracts the cpp-xlang directory to its own target, and reduces the number of queries (keep only the queries related to idl).

Another motivation for this split is to eventually generate the c++ code directly from the test thrift file. It'll be easier with a dedicated target

Reviewed By: iamirzhan

Differential Revision: D60233939

fbshipit-source-id: 1530bd1a99a2a25b166ae145d1d7fc435b600de5