Print the name of syscalls
Summary:
This will make the debug output easier to read, by printing the name of the syscall instead of the generic CallFun_*_* instruction.

It does obscure things a little, because the actual instruction is still the CallFun_*_* instruction, but I think this makes it much easier to understand what's going on in the bytecode.

Reviewed By: malanka

Differential Revision: D59520284

fbshipit-source-id: efe234bd4fbad09ac3f1d49a58ff25829c292d16