index enums
Summary:
Extend HackToThrift coverage with enum and enumerators.

```
enum CountOnlyMode {  // Enum declaration
  Disabled = 0,       // Enumerator "Disabled" ...
  HitCount = 1,
  FileCount = 2,
}
```

This works similarly to struct/union/exception types and fields.

When indexing an enum declaration, we also generate a thrift declaration fact, and remember the enum declaration in the  mutable `thrift_context`.

When indexing an enumerator, we access the enum declaration from the thrift context, and  we the enumerator identifier, we have all the information needed to generated the thrift xref.

We don't need comment annotations in the case of enumerating since the enumerator name mapping is the identity.

Reviewed By: josefs

Differential Revision: D60450750

fbshipit-source-id: a81f49bd87b5526b14ebccca2522d0b4be4a9de7