Make restorable dbs restorable
Summary:
We now wait in two phases to make sure that the restore happens correctly. In the first phase we wait to ensure that the restore process is started. In the second stage we wait for it to finish.

Previously, we didn't wait for it to start so we would get an error message.

Reviewed By: donsbot

Differential Revision: D58786113

fbshipit-source-id: fff3e03f2a14b88caa5f048460ef9c53ac864516