Add enumerator support for CxxToThrift.
Summary:
The codegen from Thrift to C++ is quite straight forward, where the enumerators are always named consistently:
```
enum CountOnlyMode {
  Disabled = 0,
  HitCount = 1,
  FileCount = 2,
}
```
generates
```
/** Glean {"file": "test.thrift", "name": "CountOnlyMode", "kind": "enum" } */
enum class CountOnlyMode {
  Disabled = 0,
  HitCount = 1,
  FileCount = 2,
};
```
To produce the `Fbthrift::EnumValue` fact, we need the `Fbthrift::NamedType` of the `enum` declaration the enumerator belongs to and the name of the enumerator. Since the enumerator name is the same between Thrift and C++, we simply use the C++ identifier without needing to rely on reading a comment.

The `Fbthrift::NamedType` instance is stashed in the `EnumDecl` object as a special case, since it will be needed upon visiting the enumerators.

Reviewed By: phlalx

Differential Revision: D60707938

fbshipit-source-id: eb7d778b9b94548390d4f7c7a2793ea2f49e964b