Only copy large binary files
Summary: If an input to `glean merge` is already larger than the specified max merged file size, it will just copy it to the output. If the input is in json format, then this causes merge to output json which is unexpected. Only perform the copy if the input is in binary format.

Reviewed By: malanka

Differential Revision: D60988559

fbshipit-source-id: a8e1ffbe1d152ace95def3884d1d53896d016bef