Disable non-deterministic symbol id test for now
Summary:
This test catches a non-deterministic behavior due to duplicate symbol ids for object and types in flow.

T198806238

To fix we should properly disambiguate typeDecl and object , or else remove the source of the duplicate entity. Given the spans are identical i'm not sure its useful having both an obejct and a type for the same span for a flow class.

For now, disable the non-deterministic part of the test so that we get a green CI.

these are the non-determinstic results https://github.com/facebookincubator/Glean/actions/runs/10367245518

that depend on the ordering of which fact gets produced first I think.

Reviewed By: malanka

Differential Revision: D61253333

fbshipit-source-id: 95b3c79d97fa7e5c4d723762669d973a16ee5ed1