Update tests to catch ambiguity between typeDecl and localDecl entities for `class` syntax
Summary:
The new module parent tests were behaving non-deterministically.
They exposed that import.js has a `class CazClass` line that is both a typeDecl and a localDecl (i.e an object and a flow type ?)
With the same symbol id generated, this would randomly insert the description of one or the other in the test output description map.

Start on the fix by correcting a bunch of the flow type errors in the test suite and modernizing the glass tests.
Add extra tests for the module member cases so we can reliably catch this situation in multiple places.

We have at least two issues lurking:
- duplicate typeDecl/memberDecl for a class
- inconsistent indexing of .js.flow vs .js files wrt. their short string names

The latter means sometimes we have the .js file for the string name, and in other cases the .js.flow fact. (see e..g flow.FileOfStringModule for the actual file test/es_exports.js.flow not having any facts). There's bugs lurking in the .js.flow suffix handling.

Reviewed By: malanka

Differential Revision: D61253172

fbshipit-source-id: 614873c8872de9436bfad9e96622df1545a03732