index thrift fields
Summary:
There's a bit of redundancy in the generated thrift annotations.

```
/** Glean {"file": "test.thrift", "name": "GetNavigationRequest", "kind": "struct" } */
class GetNavigationRequest {

  public:
   /** Glean {"file": "test.thrift", "name": "GetNavigationRequest", "kind": "struct", "field": "identifier" } */
   int get_identifier();
};
```

This made it easier for the indexer, but large generated output can be a problem (see D60844431).

This diff updates the clang indexer so it can deal with just the "field" key, and get the remaining data from the enclosing struct.

## follow up
- I'd like to move all the thrift logic in ast.cpp to a separate file for better separation of concerns.

Reviewed By: donsbot

Differential Revision: D61028619

fbshipit-source-id: b2f61adaff797876c89c38d2aa2c36a4e3d34da0