Move CC dcap query for Hack containers into predicate
Summary:
Rather than building up a multi-step query on the server, and testing the trait/interface and class cases separately, let's just push this into the server and do it there.

Benefit:
- less code to maintain on the client
- easier to test
- performance should be better as we make 1 call instead of 3, for tiny amount of extra branching on server
- easier to identify in logs

In particular, I want to do this to do disjuctions of mulitple symbols at once, instead of one call per symbol in the file.

Reviewed By: phlalx

Differential Revision: D61843558

fbshipit-source-id: ed04010ee3dd7d4d1c6a15313cf6922aecc8909f