Infer the left-hand-side of dot
Summary:
This example should work:

```
cxx1.RecordDeclaration R where R.name.name = "vector"
```

but it currently fails because the type of R is not known at the time
we typecheck `R.name`.  This diff fills in the missing inference
functionality. It required adding the ability to "demote" an
expression from a predicate type to its key type, which is the dual of
the "promote" operator we already had.

With this change I'm happy enough with dot syntax that I think we can
advertise it more widely.

Reviewed By: josefs

Differential Revision: D62499221

fbshipit-source-id: c1eec9af52457c9b246d3794df3a501955333ec4