complete migration to TypeInfo
Summary:
There are two type predicates in the Hack schema.

The legacy, deprecated, `type`, and the most recent `TypeInfo`.

`TypeInfo` contains the pretty-printed type, but also xrefs relative to that string. Codehub uses it to generate go-to-def links the API doc page.

Not all definitions contain a TypeInfo field yet. This diff updates the schema to add it everywhere.

Motivation:
 - consistency
 - if we want to index the Hack type AST, it'll be simpler to add a new field in TypeInfo, which is now available everywhere (see linked task)

Reviewed By: pepeiborra, donsbot

Differential Revision: D62471296

fbshipit-source-id: 17bf41cab000600d294058b9a3c4293c4504ecd2