Simplify the way the schema ID is supplied to a query
Summary:
We had some legacy cruft here, and things were more confusing than
necessary.

The new story is:
* The schema ID for a query is always specified by the `schema_id`
field of `UserQuery` (or `UserQueryFacts`). The field is populated
by the client.
* In the local backend, `envSchemaId` stores the schema ID for the
session, and this is attached to each query in `instance Backend
Env`.
(previously it was attached in `Glean.Query.Thrift`, which
missed some queries, but it was also being picked up inside
`userQuery` itself which was pretty confusing).
* The `--schema-id` flag is now ignored. I don't think there's a good
use for this, we should always be passing the schema ID with the query.
* We now log a warning if the user-supplied schema ID is unknown. This
might help debug some issues like D62471296

Reviewed By: josefs

Differential Revision: D62737521

fbshipit-source-id: e7bddaf3c5e9b11dd9aa8d617cb0827e74e002da