Fixes for MatchArrayPrefix
Summary:
We had a bug where if the optimiser substituted an array prefix
pattern, we could get a runtime crash from `error "buildTerm"` because
it doesn't know how to reconstruct the array that matched the prefix.

All patterns have to support `buildTerm`, so we need to augment
`MatchArrayPrefix` with either a pattern that matches the whole array,
or a pattern that matches the array suffix. The former is much easier,
so that's what I've done.

This exposed a couple of other bugs in the query engine related to how
we compile `QueryAnd` and `QueryArrayPrefix`, so I've fixed those.

Reviewed By: josefs

Differential Revision: D62871016

fbshipit-source-id: 19a9393b392f3959a591737306c13d5bf4f7d4cf