Fix bug in reordering choice statements
Summary:
If a variable is bound in a subset of the branches of a choice, that's
a binding error. Previously we just renamed these variables and made
them local, but that's wrong (see the test case).

The reordering pass will catch the binding error and attempt a
different order, so we'll be able to discover a valid ordering if one
exists.

Reviewed By: donsbot

Differential Revision: D63323890

fbshipit-source-id: 8afe449bf38edad225d27f3634911374f5990ce5