Convert up to 11 function arguments for hack.FileCall
Summary:
FileCall arguments must be translated from indexer-/language-specific
predicates to codemarkup.FileCall.  Sadly, this means rigid pattern
matches of the argument list, so we lose FileCalls if they have too
many arguments.  This recently happened for a set of calls our partner
teams care deeply about.

Python arg translation goes up to 11, so I've adjusted Hack to match.

I had hoped to make the last clause something like "11 or more" using
pepeiborra's [A,B,C,..] match notation, but it was not to be -
something in the translation or the underlying query engine really
doesn't like this, and the resulting clauses only match if I add a
second array match clause of the right length in just the right place.

Reviewed By: phlalx

Differential Revision: D63483508

fbshipit-source-id: 8ed84f53239f3ef3e359b1220c69556b864a64cc