Fix a deriving bug
Summary:
This fixes a bug that can result in invalid facts being derived.  See the Note in `UserQuery.hs` for full details

The test currently doesn't fail for accidental reasons, but the bug
was exposed by the upcoming changes in D62647670. Therefore the fix is
needed before I can land that.

Reviewed By: malanka

Differential Revision: D63540663

fbshipit-source-id: c0bb0d559bda851a2cad6bd44d7bdd656e23d778