Show detailed component information on hover for component declarations
Summary:
This diff reuses the same infra added in D62917950, to show the same detailed information for component declarations as well.

I noticed in the re-recorded tests that we are printing out a lot of `renders React.Node` for component without render declaration. This will be addressed separately in the next diff.

Changelog: [ide] Hovering on components will now consistently show its props and renders information. Previously, the information is omitted for component declarations.

Reviewed By: panagosg7

Differential Revision: D63660744

fbshipit-source-id: dc049025d805dfc4b627ed3f4dc7e23956d36533