fetchSymbolsAndAttributesGlean can query a different backend
Summary:
When computing snapshots, we need all dependent files to be indexed in order to have the xrefs. This is true for xlang navigation as well. In order to get the thrift xrefs from hack/c++, we would need to index thrift files as well.

At diff-time we index  different languages in different workflows, and the API doesn't allow to run both hack/thrift in the same workflow at the moment.

Another solution is for the snapshot generator to query an external db.  This is simpler to implement and also won't occur extra latency.

Downside is that if a diff adds thrift entities and hack/cpp client code, we may miss some xlang xrefs since the thrift symbol won't exist in the thrift db and we won't be able to resolve their location. I think it's a rare occurrence.

This diffs adds the ability for the document symbol handler to query a different backend for resolving xlang refs. This new path isn't used in Glass, but is exercised by the snapshot tool.

Reviewed By: simonmar

Differential Revision: D63459934

fbshipit-source-id: 4e588fde9a4d4f26c5efc35450c660c3e21c1e22