Code generation for sets in various langauges
Summary:
The only thing that's controversial here is that we encode sets as lists in OCaml. Generating code for sets in OCaml is a real pain because each form of element type requires a new instantiation of the `Set.Make` functor, i.e. a new module.
After discussing with mheiber we concluded that producing lists without duplicates is the most ergonomical for all parties.

Reviewed By: simonmar

Differential Revision: D63611594

fbshipit-source-id: 5bfab7402ac7c442be6fa55db07a42eb574f2360