Relaxed reordering
Summary:
This is a change to Angle optimisation to give the optimiser more flexibility when reordering statements. The goal is to move towards the Angle compiler getting smarter and doing more of the work and the human doing less, which should mean optimising Angle queries will get easier.

The current behaviour is

* Explicit statements (`A = B`) are ordered with respect to other statements in the same `where` block
* Implicit statements (e.g. those arising from nested patterns) can be reordered with respect to other implicit statements arising from the same statement

The new behaviour is

* Explicit statements (`A = B`) are ordered with respect to other statements in the same `where` block
* Implicit statements can be reordered with respect to **other statements in the enclosing `where` block**

In both cases, further reordering may take place when a statement is unresolved, or the compiler can prove it is trivial or O(1).

For example, if you write

```
F = cxx1.FunctionDeclaration { name = { name = { name = "fun" }}};
R = cxx1.RecordDeclaration { name = { name = "class" }};
```

we can now reorder the nested `{ name = "class" }` match before the `cxx1.FunctionDeclaration` match, whereas we couldn't do that previously.

The main internal change is that instead of a `FlatStatementGroup` being either fully ordered or fully unordered, we now have groups that consist of ordered statements and floating statements.

There are a few differences in perf tests. Quite a few tests improved, some got (slightly) worse - I've checked all the ones that regressed and I'm comfortable there's nothing bad going on there.

In order to get all the tests to pass, I had to make a lot of improvements to the optimiser, some of which have already landed in previous diffs and the rest of which are here.  This has been quite a long journey...

Reviewed By: josefs

Differential Revision: D64049923

fbshipit-source-id: 0a3516b9e2941ee2c448eabc998c496689425400