Allow passing directory as input
Summary:
**Why?**
fbsource cxx skycastle indexing workflow takes about 20 hours to complete with ownership and the bottleneck is writing.
https://www.internalfb.com/sandcastle/workflow/1747396655424399432
In that workflow we index targets in batches with size 3072 and merge each batch by chunks of 1024. We could try to merge more to deduplicate more and improve writing speed, but having a big chunk causes merging to be slow (not improving the total time).

Another way to merge more is to apply another merge command to all merged chunks. This way we will merge whole batch with similar performance.

It's implemented in the next diff, but for this we need `glean merge` to accept directories as an input to be passed from bxl script which is implemented here

Reviewed By: malanka

Differential Revision: D64594579

fbshipit-source-id: 2cf49576864a44fe702e3aef4eec657bf2510c78