With `:pager on`, keep fetching results
Summary:
If we're using a terminal pager, just page more results from the
server. There will be natural flow control due to the pipe between the
shell and the pager.

The exception is if the timeout is hit: I'm currently stopping the
query with the timeout message rather than continuing the query. Often
I want to get some diagnostics (e.g. `:profile full`) in the event of
a timeout, and if the user has to hit `^C` to interrupt the query we
wouldn't see these.

Reviewed By: iamirzhan

Differential Revision: D65332020

fbshipit-source-id: 908035225a59c3fc2355fc6bee2d475cd6d186a5