delete model tests
Summary: The model tests were an attempt to add model-based testing to the Glean janitor, a notoriously tricky component of Glean that is responsible for various db lifetime tasks in a Glean server. I ran out of time and energy before fully modelling it, and the tests have been flaky under stress mode, so they are pretty much dead code right now.

Reviewed By: phlalx

Differential Revision: D65272158

fbshipit-source-id: 3034c1194b19d6bb2734be0a56e6e0d5a46b2141