Code generation for concrete sets and typechecking sets
Summary:
The trickies aspect here is that we need to put all elements back in a set because the renaming of fact IDs can cause things that were previously different to become equal. So the typecheck bytecode needs access to all the set syscalls.

I have moved out the syscalls for set in their own struct which is inherited by the different invocations of the bytecode engine, for both typechecking and running the code.

Reviewed By: simonmar

Differential Revision: D63632783

fbshipit-source-id: a48a70dbceaa1fece3562d2b976013cea899981c