Support testing perf of derived predicates
Summary:
We want regression tests for the performance of derivations. We can
query a stored derived predicate as if it was a non-stored derived
predicate by setting `store_derived_facts = True` in
`UserQueryOptions`, but this only works on a writable DB.

I'm relaxing that behaviour to allow `store_derived_facts` on a
read-only DB provided `omit_results = False` (`glean derive` sets this
to `True`, so `glean derive` will still fail on a read-only DB, which
is what we want).

This allows the regression test framework to test the perf of derived
predicates, I've enabled one such test for `cxx1.DeclByName` as an
example.

Reviewed By: malanka

Differential Revision: D65533801

fbshipit-source-id: 3886f3ebfb5d714754fed8153859beaa148c567b