Refactor ordering
Summary:
Changing the representation of groups from

```
data FlatStatementGroup =
  FlatStatementGroup
    [FlatStatement] -- ordered
    [FlatStatement] -- floating
```

to

```
newtype FlatStatementGroup = FlatStatementGroup [Ordered FlatStatement]

data Ordered s
  = Ordered s
  | Floating s
```

The advantage of the latter is that we can reorder a group multiple times without losing the `Ordered` status on the statements, this is used in `reorderStmtGroup`. The new representation is also simpler to work with.

I'm still not happy with the state of all this code, but I think this is a step forwards.

Reviewed By: josefs

Differential Revision: D64897648

fbshipit-source-id: 8a683d8156d7e3101fe0578deb560018335f9746