translate repo/path in documentsSymbols() results for fbsource/www files
Summary:
This diff updates handler for documentSymbols() so it uses the same repo/paths as the queried document in the case of fbsource/www.  This is when using the Glean backend. This was done for XDB in D65616096.

For instance, if querying `fbsource/www/flib/intern/glean/GleanGlass.php`, we will have xrefs of the form `fbsource/www`, rather than `www/....`.  See D65616096 for motivation.

Reviewed By: pepeiborra

Differential Revision: D65692320

fbshipit-source-id: 9489b680bd82f4a37de0b810efb6cd01d9571fef