add mirror rev option
Summary:
Add option --mirror-revision to glass-snapshot

Generally, this is to handle the case when a repo is duplicated in another. Only use case at the moment is to address the mega repo migration: www becomes a part of fbsource. We want to add two entries to xdb for a given path in that case, to cover the two possible client queries.

Example of invocation
```
 glass_snapshot --revision REV --db-root DB_ROOT --db DB --snapshot-tier XDB FILE --mirror-revision MIRROR_REV
```
where `FILE = REPO/PATH`

Effects of this option:
- as without the option --mirror-revision option, we create a snapshot for revision REV, store it in XDB at key: (REPO, PATH, REV).
- the option also creates an *equivalent* translated snapshot at key (REPO', PATH', MIRROR_REV)

For simplicity, REPO' and PATH' aren't provided in the CLI but computed by glass-snapshot because they can be deduced from REPO, PATH.

For instance, if REPO=www, PATH=foo.php, we'll have REPO'=fbsource, PATH'=www/foo.php.

What do I mean by an  *equivalent* snapshot?

we can't reuse the exact same snapshot. At the very least, we want the snapshot revision [here](https://www.internalfb.com/code/fbsource/[c2e0400cab3f]/fbcode/glean/glass/if/glass.thrift?lines=254) to match the xdb key, because client expect them to match.

But we also want to translate the [xref targets](https://www.internalfb.com/code/fbsource/[c2e0400cab3f]/fbcode/glean/glass/if/glass.thrift?lines=185) in the mirror snapshot to use the mirror repo and paths. We keep the same symbol ids though because Glass doesn't handle fbsource/www symbols ids, and it shouldn't be an issue on the client side.

Not performing the repo/path translation causes issues:

Phabricator: internal navigation doesn't work because Phab can't reconcile those www/..  paths to the fbsource paths in the diff
Codehub: when jumping cross-repo, Codehub loses the revision of the source page

This can be fixed but it seems simpler to do it in one place.

Reviewed By: pepeiborra

Differential Revision: D65616096

fbshipit-source-id: 59c7688639cb5ccbcccefa5241860c4fefe3c31d