Schema evolution for sets
Summary:
This diff provides the bare minimum schema evolution to allow set types to be used in production.

My plan is to provide much more elaborate evolution in future diffs. But because sets are implemented as syscalls, and the schema machinery currently doesn't use syscalls, such a change is pretty involved.

Reviewed By: donsbot, simonmar

Differential Revision: D65896602

fbshipit-source-id: da82f01a1febbaf6f9a58432741ee066e5fa2590