Use lists to encode sets in Thrift
Summary:
The set type in Hack is very limited and cannot have e.g. predicates as elements. Since the thrift set type translates to the Hack set type, that means that we cannot use sets in thrift to encode sets in Angle.

This is really sad because it means that we cannot use sets in other language APIs, like Haskell :-(

Reviewed By: donsbot

Differential Revision: D66158787

fbshipit-source-id: 05beed53d0e5de40c185848b6eae16c62024354c