Uses FileXRefsGenericEntities to compute xrefs
Summary:
Use new predicates introduced in D66079681 to compute xrefs.

This allows generalization of  the computation of  xlang xrefs.

Before, glass was relying on the assumption that xlang references where the result of a mapping "generated entity <-> idlEntity".

In this diff, glass uses FileXRefsGenericEntities to compute all xrefs (plain or xlang). The xlang entities may be generated from a entity <-> idlEntity (as is the case for Hack), but this isn't mandatory. For Hack it's done in Codemarkup.

Note that the C++ handler is unchanged and still relies on the "entity <-> idlEntity" mapping.

Reviewed By: nhawkes

Differential Revision: D66128389

fbshipit-source-id: 8210052f0af20a4b8a6dbe4f939b27d380c43e5e