Use kind tables for scope search
Summary:
This is more win, as we use restricted kind tables to construct scope terms.

Note PythonNameCase, it’s a bit subtle. we want to check that the string is in one of the scope clases before building an SName. We want if-then-else with short circuiting, not disjunction. Maybe I need to use the double negation existence check trick?

Next: remove the old search deriving infastructure methodlically and write a few more tests.

Reviewed By: phlalx

Differential Revision: D66532608

fbshipit-source-id: 15ddf401228039879df04d21cc0efbca7c135a58