use correct repo for xlang xrefs for mirrored repos
Summary:
Reminder: fbsource/www and www/ paths both map to the same dbs. In Glass, we consider that "fbsource/www" is a mirror of www, and we're careful to translate the returned paths so they match the repo query. See this [post](https://fb.workplace.com/groups/csi.eng/permalink/1990139908151741/).

There's a corner case that wasn't handled correctly: When processing DocumentSymbols() query, we need to resolve the location of cross-dbs entities. We pick the db based on
- the query repo
- the cross-db target entity language

The rules are currently
- (fbsource, thrift) -> fbsource.fbthrift
- (www, thrift) -> www.fbthrift

There's an issue in the case of fbsource/www hack documents. The source repo is fbsource, which means we use the wrong thrift db.

Proposed fix is to again use the MirrorConfig mapping and make sure we use the "origin" repo to determine the target entity db (e.g. www instead of fbsource).

Reviewed By: iamirzhan

Differential Revision: D66679488

fbshipit-source-id: a640830d21c5ee9eafc6dd3cc6eb32ad55e038db