Release fact batches earlier
Summary:
When we added concurrent commit, it was hard to also release the
committed batch as we had been doing previously because it might be in
use by the next rename. Here I'm adding an explicit release, but it
requires the write lock to avoid race conditions; we'll have to see
how well this works in practice.

Reviewed By: iamirzhan

Differential Revision: D66695438

fbshipit-source-id: ff400ebad934bace2919723ba7f3caae24e1fe45