introduce "generates" relation type for searchRelated method
Summary:
leverage the codemarkup predicate introduced in D66787155.

Note: unlike the other relation type, we don't compute the symbol locations and details. This isn't mandated by the service API. Could be done in the future, by reusing the logic we use to resolve location of xlang xrefs

motivation: compute some xlang xrefs on the fly when we don't want to rely on DocumentSymbols(), for instance in VSCode. This could be easier to integrate with the LSP. Discussed [here](https://fb.workplace.com/groups/937031054788533/permalink/1064818692009768/), also see test plan in D66792443

Reviewed By: donsbot

Differential Revision: D66794767

fbshipit-source-id: 088a66a94c487606b48ffe812df2b5793bdec8d2