Release ownership memory earlier
Summary:
After substituting an ownership vector we had to wait for the GC to
run the finalizer to free the old vector, which can cause a space
leak. This releases the old immediately. which is unsafe but we're
using it safely.

Also moved these operations into IO which avoids the need for explicit
forcing.

Reviewed By: iamirzhan

Differential Revision: D67027268

fbshipit-source-id: c48da3f45c647babbbc832229333cd858353e58d