Add a dereference operator, .*
Summary:
When using dot syntax it's sometimes useful to be able to refer to
"the key of this fact", and while we often do the dereference
implicitly as in `X.field` when `X` is a predicate type, there's no
way in general to just select the key from a predicate. In dot syntax
this leads to sometimes getting stuck and having to go back to pattern
matching to extract a key.

This diff adds a dereference operation `.*` for selecting the key from
a fact. Typechecking it requires a new constraint-like construct
`HasKey` which expresses the constraint that a type is a predicate
with a given key type.

Reviewed By: donsbot

Differential Revision: D67102320

fbshipit-source-id: 4f6f445950d11108fd402bd243e79f029fe83bc0