Schema evolution for sets
Summary:
This diff enables two things:
* Schema evolution for elements of sets.
* Schema evolution from array to set. This is very handy for starting to use sets in place of arrays.

Since sets are implemented using so called "syscalls" they are required when evolve and transform terms. The code that does the transformation didn't have access to the syscalls before, and it is too clever for its own good. So there's a lot of noise in this diff to pass along the syscalls.

Reviewed By: simonmar

Differential Revision: D67345466

fbshipit-source-id: 04aa59bced097572426354cad559ff2fb05a1ee2