Use set for the collection of xrefs in a file
Summary:
The collection of xrefs in a file is really a set since it's unordered. We've had problems in the past when the Hack indexer was non-deterministic and would produce different arrays in different runs. That's currently not a problem, but using sets here is the right thing to do.

This should be a no-op.

Reviewed By: simonmar

Differential Revision: D67763122

fbshipit-source-id: 1b4b94a632edce3546b5095fe1ce137172cebab5