GraphQL Glean: Add schema name to BelongToConfig
Summary:
Since GraphQL types are unique per schema, we need to know the schema name for diff labeling. (Probably for IDE support / navigation tasks too, if we're being rigorous.)

This diff adds it to the GraphQL Glean schema alongside the build config name. It is a property of the build config.

The schema name is technically optional in the build config, so I propagated that through here. (It has to be optional in the Glean schema anyway for backwards compatibility.)

Reviewed By: ginfung

Differential Revision: D68475056

fbshipit-source-id: a28e19cbde9e2fc1f3a3e0778d51df659659275f